fix(fresco): address terminal restoration review
